All-link command response is always 6 bytes.
When an all-link command arrives from the IM to host, per the Modem
Developer's Guide, it's always 6 bytes. The previous behavior left an
extra hex pair on the front of the buffer.

It's possible to receive All-link commands coming from linked
responders even if you didn't originate a scene control command
yourself, if user has a scene control keypad.